test: EntityFactory의 postEntities 메소드를 DummyPost로 이동
